finish

+++ Nextion.cpp
@@ -394,7 +394,7 @@ bool Nextion::receiveNumber(uint32_t *number)
   if (temp[0] == NEX_RET_NUMBER_HEAD && temp[5] == 0xFF && temp[6] == 0xFF &&
       temp[7] == 0xFF)
   {
-    *number = (temp[4] << 24) | (temp[3] << 16) | (temp[2] << 8) | (temp[1]);
+    *number = (uint32_t(temp[4]) << 24) | (uint32_t(temp[3]) << 16) | (temp[2] << 8) | (temp[1]);
     return true;
   }
